bugs

+++ POGeneticAlgorithm/math_helper.h
@@ -108,7 +108,7 @@ const Vector2 VEC2_RIGHT = { 1.0f, 0.0f };
 
 static float distnaceBetweenVecs(Vector2 vec1, Vector2 vec2)
 {
-	return sqrtf(pow(vec2.y - vec1.y, 2) + pow(vec2.x - vec1.x, 2));
+	return sqrtf(powf(vec2.y - vec1.y, 2) + powf(vec2.x - vec1.x, 2));
 }
 
 struct BoundingBox
